+++ src/main/java/com/rodrigo/AppContatos/services/PessoaService.java
@@ -21,11 +21,37 @@ public class PessoaService {
         return pessoaRepository.findById(id);
     }
 
+    public Pessoa update (Pessoa pessoa){
+        Optional<Pessoa> pessoaJaExistente = pessoaRepository.findById(pessoa.getId());
+
+        if (pessoaJaExistente.isPresent()){
+            Pessoa updatePessoa = pessoaJaExistente.get();
+            updatePessoa.setNome(pessoa.getNome());
+            updatePessoa.setCep(pessoa.getCep());
+            updatePessoa.setCidade(pessoa.getCidade());
+            updatePessoa.setEndereco(pessoa.getEndereco());
+            updatePessoa.setUf(pessoa.getUf());
+            return pessoaRepository.save(updatePessoa);
+        }else {
+            return pessoaRepository.save(pessoa);
+        }
+    }
+
     public Pessoa save(Pessoa pessoa) {
-        return pessoaRepository.save(pessoa);
+        Optional<Pessoa> pessoaJaExistente  = pessoaRepository.findById(pessoa.getId());
+        if (pessoaJaExistente.isPresent()){
+            System.out.println("Pessoa ja Cadastrada no banco de dados");
+            return pessoaJaExistente.get();
+        }else{
+            return pessoaRepository.save(pessoa);
+        }
     }
 
     public void deleteById(Long id) {
+        Optional<Pessoa> pessoa = pessoaRepository.findById(id);
+        if (pessoa.isEmpty()){
+            throw new RuntimeException("Pessoa não encontrada");
+        }
         pessoaRepository.deleteById(id);
     }
 }
